pyproject.toml
[tool.poetry]
name = "pythondrftestforwork"
version = "0.1.0"
description = ""
authors = ["Евгений Кузнецов <eugeny.kuz@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
redis = {extras = ["hiredis"], version = "^5.0.3"}
ffmpeg-python = "^0.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}


[tool.poetry.group.production.dependencies]
gunicorn = "^22.0.0"


[tool.poetry.group.development.dependencies]
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.1.1"
flake8-bugbear = "^24.4.21"
pre-commit = "^3.7.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "drf_ffmpeg.settings"
python_files = ["test_*.py", "*_test.py"]

[tool.black]
target-version = ['py311']

[tool.isort]
profile = "black"
